test_cc1541: Improve runtime speed of the testsuite.
The variables named test and passed are incremented
frequently.  Since they serve the purpose of a simple
counter, only, without any side-effects, we should prefer
the pre-increment operator on them.

When the return value is ignored, the "pre" form (++i)
is never less efficient than the "post" form (i++), and
is often more efficient.  This is because post-increment
(or decrement) requires a copy of i to be made, which is
the value of the expression.